refactor(tryCount): tryCount를 불변값인 VO 대신 DTO로 사용하도록 수정
