+++ src/main/java/bridge/service/BridgeGame.java
@@ -2,7 +2,7 @@ package bridge.service;
 
 import bridge.vo.enums.Step;
 import bridge.vo.StepResult;
-import bridge.vo.TryCount;
+import bridge.dto.TryCountDto;
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
@@ -22,7 +22,7 @@ public class BridgeGame {
      * <p>
      * 재시작을 위해 필요한 메서드의 반환 타입(return type), 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
      */
-    public TryCount retry(TryCount tryCount) {
-        return tryCount.addCount();
+    public void retry(TryCountDto tryCount) {
+        tryCount.addCount();
     }
 }
